Process lock LocalBuildpack
If a LocalBuildpack instance is shared between multiple threads or processes it will attempt to produce an image with the same name but it will error out. We can prevent that by generating a temp file and taking out a lock on it. This lock is visible across all processes and is atomic. Locking a file in this way prevents multiple buildpacks with the same name from being built by pack at the same time.