Allows exercising Salesforce Functions (#10)
* Allows exercising Salesforce Functions

This PR adds a FunctionQuery class that enables a running container (via `App#start_container`) that contains a salesforce function to be exercised.

This is an experimental API and may change in patch versions. See spec/unit/function_query_spec.rb for example usage.

* Webmock function query spec

To be able to iterate on the FunctionQuery class rapidly this commit adds web mock. We can assert body, headers, and port against requests made by Excon which is wrapped in the FunctionQuery class.

* Clean up Java test app

- Remove unused testing file
- Reduce dependencies

* Update JVM function repo

* Update JVM function test

Co-authored-by: Manuel Fuchs <manuel.fuchs@salesforce.com>