Prefer local buildpack over docker image
Packaging up the local buildpack into a docker image is an unnecessary step as it's just taking the files on disk and sticking them in a docker image. This process is slower than giving `pack` the location of the files directly.

Generating a docker image also introduces cleanup problems. When multiple processes use the same image name, the first process to end will delete the docker image even though other tests are still using it.

This removes putting buildpacks into docker and uses the files on disk directly.