Introduce "exactly once" semantics into LocalBuildpack `build.sh` builds
There was already an unsuccessful attempt to add mutex coordination in LocalBuildpack. It didn’t work because it used a tempfile and the instances are created in each process rather than being forked. If we moved the file lock into a common directory then we could avoid that problem and introduce process coordination.

For this implementation to work we have to have exactly-once semantics. The target dir must be deleted once, and the build.sh must be executed once per test run execution. To that end, we must be able to detect when a test run is triggered and whether or not a process has already performed the delete and build actions.

This is done by generating a SHA of the current buildpack on disk, then storing that value in the locked file. When the contents on the disk change the SHA will also change and the build.sh can be re-run. 

https://salesforce.quip.com/veWYASMAibmN